refactor: use node-recommended way for dual packages
