+++ fetch.ts
@@ -3,7 +3,7 @@ import rdf from 'rdf-ext'
 import formats from '@rdfjs/formats-common'
 import rdfFetch, { FactoryInit, DatasetResponse } from '@rdfjs/fetch-lite'
 import { RdfXmlParser } from 'rdfxml-streaming-parser'
-import { expand, loadFile, prefixes } from './src/lib'
+import { expand, loadFile, prefixes } from './src'
 import { NamedNode } from 'rdf-js'
 import DatasetExt from 'rdf-ext/lib/Dataset'
 import { overrides, Override } from './overrides'
