+++ package.json
@@ -2,10 +2,14 @@
   "name": "@zazuko/rdf-vocabularies",
   "version": "2020.6.17",
   "description": "Access most commonly used RDF ontologies/schemas/vocabularies as datasets or n-quads",
-  "main": "lib/node/index.js",
-  "module": "lib/es/index.js",
+  "main": "lib/index.js",
+  "module": "lib/index.mjs",
   "bin": {
     "rdf-vocab": "bin/vocab.js"
+  },
+  "exports": {
+    "require": "lib/index.js",
+    "import": "lib/index.mjs"
   },
   "files": [
     "bin",
@@ -15,15 +19,14 @@
     "*.d.ts.map"
   ],
   "scripts": {
+    "prefetch": "mkdirp ontologies",
     "fetch": "ts-node fetch.ts",
     "lint": "eslint . --ext .js,.ts",
     "pretest": "npm run lint",
     "test": "jest",
-    "build:node": "tsc --target es2017 --outDir lib/node",
-    "build:modern": "tsc --module esnext --outDir lib/es",
-    "build:types": "tsc --outDir . --declaration --declarationMap --emitDeclarationOnly",
-    "build:es-loader": "tsc src/es/loadDataset.ts --target esnext --module esnext --esModuleInterop --moduleResolution node --outDir lib/es",
-    "build": "rimraf lib; npm run build:node; npm run build:modern; npm run build:es-loader; npm run build:types",
+    "build:node": "tsc",
+    "build:modern": "babel src --out-dir lib --extensions .ts --out-file-extension .mjs --ignore *.test.js --ignore src/loadDataset/index.ts",
+    "build": "run-p build:*",
     "prepack": "npm run build"
   },
   "repository": {
@@ -37,6 +40,9 @@
   },
   "homepage": "https://github.com/zazuko/rdf-vocabularies#readme",
   "devDependencies": {
+    "@babel/cli": "^7.10.3",
+    "@babel/preset-env": "^7.10.3",
+    "@babel/preset-typescript": "^7.10.1",
     "@rdfjs/fetch-lite": "^2.0.0",
     "@rdfjs/formats-common": "^2.1.0",
     "@types/jest": "^24.9.1",
@@ -56,6 +62,7 @@
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.0",
     "jest": "^24.8.0",
+    "npm-run-all": "^4.1.5",
     "rdfxml-streaming-parser": "^1.3.3",
     "rimraf": "^3.0.2",
     "ts-jest": "^24.3.0",
